sdxl/sdxl-studio/runtime/requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu124/
accelerate==1.3.0; python_full_version >= '3.9.0'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
diffusers==0.32.2; python_full_version >= '3.8.0'
fastapi==0.115.8; python_version >= '3.8'
filelock==3.17.0; python_version >= '3.9'
fsspec==2025.2.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.4; python_full_version >= '3.8.0'
huggingface-hub==0.28.1; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.6.1; python_version >= '3.9'
jinja2==3.1.5; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
mpmath==1.3.0
networkx==3.4.2; python_version >= '3.10'
numpy==2.2.2; python_version >= '3.10'
nvidia-cublas-cu12==12.4.5.8; python_version >= '3'
nvidia-cuda-cupti-cu12==12.4.127; python_version >= '3'
nvidia-cuda-nvrtc-cu12==12.4.127; python_version >= '3'
nvidia-cuda-runtime-cu12==12.4.127; python_version >= '3'
nvidia-cudnn-cu12==9.1.0.70; python_version >= '3'
nvidia-cufft-cu12==11.2.1.3; python_version >= '3'
nvidia-curand-cu12==10.3.5.147; python_version >= '3'
nvidia-cusolver-cu12==11.6.1.9; python_version >= '3'
nvidia-cusparse-cu12==12.3.1.170; python_version >= '3'
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5; python_version >= '3'
nvidia-nvjitlink-cu12==12.4.127; python_version >= '3'
nvidia-nvtx-cu12==12.4.127; python_version >= '3'
nvitop==1.4.2; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pillow==11.1.0; python_version >= '3.9'
pip==25.0; python_version >= '3.8'
psutil==6.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.7.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
safetensors==0.5.2; python_version >= '3.7'
setuptools==75.8.0; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.45.3; python_version >= '3.9'
sympy==1.13.1; python_version >= '3.8'
tokenizers==0.19.1; python_version >= '3.7'
torch==2.5.1+cu124
tqdm==4.67.1; python_version >= '3.7'
transformers==4.43.4; python_full_version >= '3.8.0'
triton==3.1.0; platform_system == 'Linux' and platform_machine == 'x86_64' and python_version < '3.13'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
uvicorn[standard]==0.34.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.0.4; python_version >= '3.9'
websockets==14.2; python_version >= '3.9'
xformers==0.0.29.post1; python_version >= '3.9'
zipp==3.21.0; python_version >= '3.9'
